data_app/requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
celery
redis
plotly
openpyxl